Remove locale-specific floating point from io
Use only C-locale functions wxString::ToCDouble and fmt::format{}
functions so that we do not need to set the locale for file io
operations

This also introduces a cmake check that looks for certain banned
functions (strtod, strtof and atof) that convert strings into doubles
by the user locale or formatting specifiers like %f/%g that convert
doubles into strings by the users locale.  It will prevent compilation
in this case.

For the limited cases where we want to show the user a string in their
locale, we have an override string "format:allow" that can be added to
lines in order to allow them with %f/%g format specifiers

+++ eeschema/sch_io/kicad_legacy/sch_io_kicad_legacy_lib_cache.cpp
@@ -1856,8 +1856,8 @@ void SCH_IO_KICAD_LEGACY_LIB_CACHE::saveText( const SCH_TEXT* aText, OUTPUTFORMA
         text = wxT( "\"" ) + text + wxT( "\"" );
     }
 
-    aFormatter.Print( 0, "T %g %d %d %d %d %d %d %s",
-                      (double) aText->GetTextAngle().AsTenthsOfADegree(),
+    aFormatter.Print( 0, "T %d %d %d %d %d %d %d %s",
+                      aText->GetTextAngle().AsTenthsOfADegree(),
                       schIUScale.IUToMils( aText->GetTextPos().x ),
                       schIUScale.IUToMils( -aText->GetTextPos().y ),
                       schIUScale.IUToMils( aText->GetTextWidth() ),
